reorg code into core, edit and angular pieces
