+++ bower.json
@@ -24,7 +24,8 @@
     "angular": "1.2.21",
     "angular-bootstrap": "0.11.2",
     "angular-bootstrap-colorpicker": "~3.0.8",
-    "jsonix": "highsource/jsonix#e285e82fb15fa3d62665a35eb728022936df1f4b"
+    "jsonix": "highsource/jsonix#e285e82fb15fa3d62665a35eb728022936df1f4b",
+    "ogc-schemas": "highsource/ogc-schemas#master"
   },
   "devDependencies": {
     "bootstrap": "3.0.0",
