+++ examples/style-example.html
@@ -5,7 +5,7 @@
         <meta charset="UTF-8">
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
         <link rel="stylesheet" href="../bower_components/bootstrap/dist/css/bootstrap.min.css" type="text/css">
-        <link rel="stylesheet" href="../dist/mapstory-style-editor.css" type="text/css">
+        <link rel="stylesheet" href="../dist/story-tools-edit.css" type="text/css">
         <link rel="stylesheet" href="../bower_components/angular-bootstrap-colorpicker/css/colorpicker.css" type="text/css">
 
         <script src="../bower_components/angular/angular.min.js" type="text/javascript"></script>
@@ -14,8 +14,11 @@
         <script type="text/javascript" src="../bower_components/angular-bootstrap/ui-bootstrap-tpls.js"></script>
         <script type="text/javascript" src="../bower_components/angular-bootstrap-colorpicker/js/bootstrap-colorpicker-module.js"></script>
         <script src="ol.js" type="text/javascript"></script>
-        <script type="text/javascript" src="../dist/mapstory-style-editor-tpls.js"></script>
-        <script type="text/javascript" src="../dist/mapstory-style-editor.js"></script>
+        <script type="text/javascript" src="../dist/story-tools-edit-tpls.js"></script>
+        <script src="../dist/story-tools-core.js" type="text/javascript"></script>
+        <script src="../dist/story-tools-edit.js" type="text/javascript"></script>
+        <script src="../dist/story-tools-core-ng.js" type="text/javascript"></script>
+        <script src="../dist/story-tools-edit-ng.js" type="text/javascript"></script>
         <script type="text/javascript" src="style-example.js"></script>
         <script src="http://localhost:35729/livereload.js"></script>
         <style type="text/css">
