+++ examples/svg-icon-test.html
@@ -1,0 +1,88 @@
+<!DOCTYPE html>
+<html ng-app="svg">
+    <head>
+        <title>style example</title>
+        <meta charset="UTF-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">
+        <link rel="stylesheet" href="../bower_components/bootstrap/dist/css/bootstrap.min.css" type="text/css">
+        <link rel="stylesheet" href="../bower_components/angular-bootstrap-colorpicker/css/colorpicker.css" type="text/css">
+
+        <script src="../bower_components/angular/angular.min.js" type="text/javascript"></script>
+        <script type="text/javascript" src="../bower_components/jquery/dist/jquery.min.js"></script>
+        <script type="text/javascript" src="../bower_components/bootstrap/dist/js/bootstrap.min.js"></script>
+        <script type="text/javascript" src="../bower_components/angular-bootstrap/ui-bootstrap-tpls.js"></script>
+        <script type="text/javascript" src="../dist/story-tools-core.js"></script>
+        <script type="text/javascript" src="../dist/story-tools-edit.js"></script>
+        <script type="text/javascript" src="../dist/story-tools-core-ng.js"></script>
+        <script type="text/javascript" src="../dist/story-tools-edit-ng.js"></script>
+        <script type="text/javascript">
+            var layer;
+            angular.module('svg', ['mapstory.styleEditor.services'])
+                .controller('svgc', function($scope, $timeout, iconCommons, stSvgIcon) {
+                    var self = this;
+                    this.updateColor = function() {
+                        $timeout(function() {
+                            self.icons = self.icons.map(function(icon) {
+                                return stSvgIcon.getImage(icon.uri, self.fill, self.stroke, true);
+                            });
+                        }, 100)
+                    }
+                    iconCommons.defaults().then(function(icons) {
+                        self.icons = icons;
+                    });
+                    this.fill = 'red';
+//                    Object.defineProperty(this, "fill", {
+//                        get: function () { return this._fill; },
+//                        set: function (val) { this._fill = val; updateColor();}
+//                    });
+                    Object.defineProperty(this, "stroke", {
+                        get: function () { return this._stroke; },
+                        set: function (val) { this._stroke = val; this.updateColor();}
+                    });
+                    Object.defineProperty(this, "size", {
+                        get: function () { return this._size; },
+                        set: function (val) { this._size = val; this.icons = angular.copy(this.icons); }
+                    });
+                    this._size = 64;
+                    this._fill = '#000';
+                    this._stroke = '#fff';
+//                    this.watch('stroke', updateColor);
+//                    this.watch('size', function() {
+//                        ;
+//                    });
+                }).factory('iconCommonsImpl', function() {
+                    return {
+                        defaults: ["golf-18.svg",
+                        "cafe-18.svg",
+                        "restaurant-18.svg",
+                        "car-18.svg",
+                        "harbor-18.svg",
+                        "hospital-18.svg",
+                        "farm-18.svg",
+                        "zoo-18.svg",
+                        "water-18.svg",
+                        "heart-18.svg",
+                        "town-hall-18.svg",
+                        "industrial-18.svg"].map(function(n) {
+                        return '/assets/style_editor/icons/' + n;
+                    })
+                };
+            });
+        </script>
+    </head>
+    <body ng-controller="svgc as c">
+        <div>
+            <label>size</label><input type="number" ng-model="c.size">
+        </div>
+        <div>
+            {{c.fill}}
+            <label>fill</label><input type="color" ng-model="c.fill" ng-change="c.updateColor()">
+            <label>stroke</label><input type="color" ng-model="c.stroke">
+        </div>
+        <div ng-repeat="icon in c.icons" style="float:left">
+            <img src="{{icon.dataURI}}" width="{{c.size}}" style="background-color:lightgray">
+        </div>
+        <div class="clearfix"></div>
+        <textarea cols="80" rows="10" id="svg" ng-model="svg"></textarea>
+    </body>
+</html>
