+++ examples/time-example.html
@@ -51,7 +51,8 @@
         <script type="text/javascript" src="../bower_components/jquery/dist/jquery.min.js"></script>
         <script type="text/javascript" src="../bower_components/bootstrap/dist/js/bootstrap.min.js"></script>
         <script src="../bower_components/nouislider/distribute/jquery.nouislider.min.js" type="text/javascript"></script>
-        <script src="../dist/time-controls.js" type="text/javascript"></script>
+        <script src="../dist/story-tools-core.js" type="text/javascript"></script>
+        <script src="../dist/story-tools-edit.js" type="text/javascript"></script>
         <script src="ol.js" type="text/javascript"></script>
         <script src="time-example.js" type="text/javascript"></script>
         <script src="http://localhost:35729/livereload.js"></script>
