+++ lib/core/time/controls.js
