+++ lib/core/time/line.js
