+++ lib/core/time/slider.js
