+++ lib/core/time/utils.js
