+++ lib/edit/index.js
@@ -1,0 +1,2 @@
+exports.WPSClassify = require('./style/WPSClassify');
+exports.SLDStyleConverter = require('./style/SLDStyleConverter');
