+++ lib/ng/core/style/module.js
@@ -1,0 +1,9 @@
+(function() {
+    'use strict';
+
+    angular.module('mapstory.styleEditor.core', [
+        'mapstory.styleEditor.ol3StyleConverter',
+        'mapstory.styleEditor.svgIcon',
+    ]);
+
+})();
