+++ lib/ng/core/style/ol3StyleConverter.js
