+++ lib/ng/core/style/svgIcon.js
