+++ lib/ng/edit/style.less
@@ -1,0 +1,1 @@
+@import "style/style.less";
