+++ lib/ng/edit/style/controllers/controllers.js
