+++ lib/ng/edit/style/directives/directives.js
@@ -11,7 +11,7 @@
                     property: "@property",
                     popover: "@popoverText"
                 },
-                templateUrl: 'widgets/' + templateUrl,
+                templateUrl: 'style/widgets/' + templateUrl,
                 link: function(scope, element, attrs) {
                     // @todo bleck
                     scope.$watch(function() {
@@ -32,7 +32,7 @@
     module.directive('styleEditor', function() {
         return {
             restrict: 'E',
-            templateUrl: 'style-editor.html',
+            templateUrl: 'style/style-editor.html',
             controller: 'styleEditorController',
             require: '?styleEditorController',
             scope: {
@@ -77,7 +77,7 @@
     module.directive('graphicEditor', function(stStyleChoices, ol3MarkRenderer, iconCommons, stSvgIcon) {
         return {
             restrict: 'E',
-            templateUrl: 'widgets/graphic-editor.html',
+            templateUrl: 'style/widgets/graphic-editor.html',
             scope: {
                 symbol: '='
             },
@@ -153,7 +153,7 @@
     module.directive('classifyEditor', function() {
         return {
             restrict: 'E',
-            templateUrl: 'widgets/classify-editor.html',
+            templateUrl: 'style/widgets/classify-editor.html',
             scope: true,
             link: function(scope, element, attrs) {
                 ['showMethod','showMaxClasses', 'showRange',
@@ -224,8 +224,8 @@
             link: function(scope, element, attrs) {
                 scope.$watch('currentEditor', function() {
                     var currentEditor = scope.currentEditor;
-                    var template = 'types/' + currentEditor.name.replace(' ', '-') + ".html";
-                    element.html($templateCache.get(template));
+                    var templateUrl = 'style/types/' + currentEditor.name.replace(' ', '-') + ".html";
+                    element.html($templateCache.get(templateUrl));
                     $compile(element.contents())(scope);
                 });
             }
@@ -235,7 +235,7 @@
     module.directive('rulesEditor', function() {
         return {
             restrict: 'E',
-            templateUrl: 'rules-editor.html'
+            templateUrl: 'style/rules-editor.html'
         };
     });
 })();
