+++ lib/ng/edit/style/module.js
