+++ lib/ng/edit/style/services/styleChoices.js
