+++ lib/ng/edit/style/services/styleRuleBuilder.js
