+++ lib/ng/edit/style/services/styleTypes.js
