+++ lib/ng/templates/style/rules-editor.html
