+++ lib/ng/templates/style/style-editor.html
