+++ lib/ng/templates/style/types/class-point.html
