+++ lib/ng/templates/style/types/graduated-point.html
