+++ lib/ng/templates/style/types/simple-line.html
