+++ lib/ng/templates/style/types/simple-point.html
