+++ lib/ng/templates/style/types/unique-point.html
