+++ lib/ng/templates/style/widgets/classify-editor.html
