+++ lib/ng/templates/style/widgets/color-editor.html
