+++ lib/ng/templates/style/widgets/graphic-editor.html
