+++ lib/ng/templates/style/widgets/label-editor.html
