+++ lib/ng/templates/style/widgets/number-editor.html
