+++ lib/ng/templates/style/widgets/stroke-editor.html
