+++ lib/ng/templates/style/widgets/symbol-editor.html
