--- lib/style/services/services.js
@@ -1,12 +1,0 @@
-(function() {
-    'use strict';
-
-    angular.module('mapstory.styleEditor.services', [
-        'mapstory.styleEditor.ol3StyleConverter',
-        'mapstory.styleEditor.styleRuleBuilder',
-        'mapstory.styleEditor.styleChoices',
-        'mapstory.styleEditor.styleTypes',
-        'mapstory.styleEditor.svgIcon',
-    ]);
-
-})();
